Getting efficient marshalling set up.
Still thinking about what abstractions are best here. Abstracting channels to
have methods is kind of tricky and maybe not a good idea...but neither is
copying data. The only other way that comes to mind is to use the structs for
each layer, with a new interface...that's the front-runner at the moment anyway.